Tidied test fixtures
