+++ tests/rtgsfit_verify_analytic/tests/conftest.py
@@ -1,4 +1,5 @@
 import subprocess
+import numpy as np
 from pathlib import Path
 import pytest
 import shutil
@@ -33,3 +34,10 @@ def run_pipeline():
         ["python", "-m", "rtgsfit_verify_analytic.replay_rtgsfit"],
         check=True
     )
+
+@pytest.fixture(scope="session")
+def output_dict(run_pipeline):
+    output_file = Path(cnst.DATA_DIR) / "output_dict.npy"
+    if not output_file.exists():
+        pytest.fail(f"{output_file} does not exist. Run test_run_pipeline first.")
+    return np.load(output_file, allow_pickle=True).item()
